Move dotnet_is_dotnet() into pe_utils
