+++ libyara/include/yara/pe_utils.h
@@ -83,6 +83,8 @@ int64_t pe_rva_to_offset(PE* pe, uint64_t rva);
 
 char* ord_lookup(char* dll, uint16_t ord);
 
+bool dotnet_is_dotnet(PE* pe);
+
 #if HAVE_LIBCRYPTO
 #include <openssl/asn1.h>
 time_t ASN1_get_time_t(const ASN1_TIME* time);
